requirements.txt
amqp==5.3.1
anyio==4.8.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
black==24.10.0
bs4==0.0.2
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.10
cron-descriptor==1.4.5
Django==5.1.5
django-celery-beat==2.7.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
flake8==7.1.1
flower==2.0.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
kombu==5.4.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
ruff==0.8.6
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tornado==6.4.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
